backend/prediction_service/requirements.txt
fastapi>=0.116.1
uvicorn>=0.35.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.43
transformers>=4.55.3
torch>=2.8.0
sentence-transformers>=5.1.2
scikit-learn>=1.7.2
pandas>=2.3.3
pydantic>=2.0.0